server/requirements.txt
requests==2.25.0
psycopg2==2.8.4
flask==1.1.2
flask-assets==2.0
flask-script==2.0.6
python-dotenv==0.15.0
flask-sqlalchemy==2.4.4
flask-migrate==2.5.3
pytest==5.4.3